+++ postchain-base/src/main/kotlin/net/postchain/gtv/gtvml/GtvMLEncoder.kt
@@ -32,7 +32,7 @@ object GtvMLEncoder {
              */
             is GtvNull -> objectFactory.createNull(null)
             is GtvString -> objectFactory.createString(gtv.string)
-            is GtvInteger -> objectFactory.createInt(BigInteger.valueOf(gtv.integer))
+            is GtvInteger -> objectFactory.createInt(gtv.integer)
             is GtvByteArray -> objectFactory.createBytea(gtv.bytearray) // See comments in GTXMLValueEncodeScalarsTest
             is GtvArray -> createArrayElement(gtv)
             is GtvDictionary -> createDictElement(gtv)
